ChannelDetail.js
created the ChannelDetailss component. Did state management with useState and did clean up with useEffect. Fetched the link(page id ) from the youtube Api in the constant folder. Imported Videos and ChannelCard components.

Updated the borderRadius and added the fallback image on the CardMedia component (demoThumbnailUrl) in the VideoCard.jsx component.